space detection

+++ NeuralNetwork/row_matrix.c
@@ -14,7 +14,7 @@ void print_matrix(int rows, int cols, double m[])
     {
         for(int j=0; j<cols; j++)
         {
-            printf("%f", m[i*cols+j]);
+            printf("%d", (int)m[i*cols+j]);
         }
         printf("\n");
     }
